tidies tray functionality into a separate file, extending the Tray class with a new sub-class
